Update nested_list.R
add a way to export as headings, default behavior not changed.
NB: may need a message if type given wrongly.

NB2: the default does order sublist elements but not first level elements, I did not change that, but may not be what people want (?). Also the ordered list way gets stange sublisting due to that.